pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rotator-cli"
version = "0.1.0"
description = "A CLI tool for automatically detecting and fixing image orientation using ML"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "mergd", email = "mergd@fldr.zip" }]
keywords = ["image", "orientation", "rotation", "cli", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "check-orientation>=0.0.3",
    "Pillow>=10.0.0",
    "click>=8.0.0",
    "tqdm>=4.62.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "albumentations>=1.0.0",
    "iglovikov-helper-functions>=0.0.53",
    "numpy>=1.20.0",
]

[project.scripts]
rotator = "rotator_cli.main:main"

[project.urls]
Homepage = "https://github.com/mergd/rotator-cli"
Repository = "https://github.com/mergd/rotator-cli"
Issues = "https://github.com/mergd/rotator-cli/issues"

[project.optional-dependencies]
dev = ["ruff>=0.1.0", "pytest>=7.0.0", "pytest-cov>=4.0.0"]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "ICN", # flake8-import-conventions
    "S",   # flake8-bandit
]
ignore = [
    "S101", # assert
    "S603", # subprocess-without-shell-equals-true
    "S607", # start-process-with-partial-path
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["rotator_cli"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
